Done with question
proceed with student examination and survey

+++ app/Http/Controllers/QuestionController.php
@@ -35,6 +35,14 @@ class QuestionController extends Controller
     }
 
     /**
+     * Display the specified resource.
+     */
+    public function see(Question $question)
+    {
+        //
+    }
+
+    /**
      * Show the form for editing the specified resource.
      */
     public function edit(Question $question)
@@ -55,11 +63,24 @@ class QuestionController extends Controller
      */
     public function destroy(Question $question)
     {
-        Storage::disk('public')->delete($question->a_file);
-        Storage::disk('public')->delete($question->b_file);
-        Storage::disk('public')->delete($question->c_file);
-        Storage::disk('public')->delete($question->d_file);
-        Storage::disk('public')->delete($question->question_file);
+        if ($question->a_file != null) {
+            Storage::disk('public')->delete($question->a_file);
+        }
+
+        if ($question->b_file != null) {
+            Storage::disk('public')->delete($question->b_file);
+        }
+
+        if ($question->c_file != null) {
+            Storage::disk('public')->delete($question->c_file);
+        }
+        if ($question->d_file != null) {
+            Storage::disk('public')->delete($question->d_file);
+        }
+        if ($question->question_file != null) {
+            Storage::disk('public')->delete($question->question_file);
+        }
+
 
         $question->delete();
 
@@ -68,6 +89,29 @@ class QuestionController extends Controller
 
     public function destroyAll(Request $request)
     {
+        // Question::destroy($request->items);
+        $question = Question::whereIn('id', $request->items)->get();
+
+        $question->each(function ($item) {
+            if ($item->a_file != null) {
+                Storage::disk('public')->delete($item->a_file);
+            }
+
+            if ($item->b_file != null) {
+                Storage::disk('public')->delete($item->b_file);
+            }
+
+            if ($item->c_file != null) {
+                Storage::disk('public')->delete($item->c_file);
+            }
+            if ($item->d_file != null) {
+                Storage::disk('public')->delete($item->d_file);
+            }
+            if ($item->question_file != null) {
+                Storage::disk('public')->delete($item->question_file);
+            }
+        });
+
         Question::destroy($request->items);
 
         return response(200);
